Rename game_adapter.py
