+++ tsumemi/src/tsumemi/kif_browser_gui.py
@@ -11,7 +11,7 @@ import tsumemi.src.tsumemi.problem_list as plist
 import tsumemi.src.tsumemi.timer as timer
 
 from tsumemi.src.tsumemi.board_canvas import BoardCanvas
-from tsumemi.src.tsumemi.game_adapter import MoveInputHandler
+from tsumemi.src.tsumemi.move_input_handler import MoveInputHandler
 from tsumemi.src.tsumemi.nav_controls import FreeModeNavControls, SpeedrunNavControls
 from tsumemi.src.tsumemi.settings_window import SettingsWindow, CONFIG_PATH
